+++ packages/web/src/app/[domain]/browse/[...path]/page.tsx
@@ -2,14 +2,15 @@ import { FileHeader } from "@/app/[domain]/components/fileHeader";
 import { TopBar } from "@/app/[domain]/components/topBar";
 import { Separator } from '@/components/ui/separator';
 import { getFileSource } from '@/features/search/fileSourceApi';
-import { isServiceError } from "@/lib/utils";
+import { cn, getCodeHostInfoForRepo, isServiceError } from "@/lib/utils";
 import { base64Decode } from "@/lib/utils";
-import { CodePreview } from "./codePreview";
 import { ErrorCode } from "@/lib/errorCodes";
 import { LuFileX2, LuBookX } from "react-icons/lu";
 import { notFound } from "next/navigation";
 import { ServiceErrorException } from "@/lib/serviceError";
 import { getRepoInfoByName } from "@/actions";
+import { CodePreviewPanel } from "./components/codePreviewPanel";
+import Image from "next/image";
 
 interface BrowsePageProps {
     params: {
@@ -49,7 +50,18 @@ export default async function BrowsePage({
     })();
 
     const repoInfo = await getRepoInfoByName(repoName, params.domain);
-    if (isServiceError(repoInfo) && repoInfo.errorCode !== ErrorCode.NOT_FOUND) {
+    if (isServiceError(repoInfo)) {
+        if (repoInfo.errorCode === ErrorCode.NOT_FOUND) {
+            return (
+                <div className="flex h-full">
+                    <div className="m-auto flex flex-col items-center gap-2">
+                        <LuBookX className="h-12 w-12 text-secondary-foreground" />
+                        <span className="font-medium text-secondary-foreground">Repository not found</span>
+                    </div>
+                </div>
+            );
+        }
+
         throw new ServiceErrorException(repoInfo);
     }
 
@@ -62,70 +74,11 @@ export default async function BrowsePage({
         )
     }
 
-    return (
-        <div className="flex flex-col h-screen">
-            <div className='sticky top-0 left-0 right-0 z-10'>
-                <TopBar
-                    defaultSearchQuery={`repo:${repoName}${revisionName ? ` rev:${revisionName}` : ''} `}
-                    domain={params.domain}
-                />
-                <Separator />
-                {!isServiceError(repoInfo) && (
-                    <>
-                        <div className="bg-accent py-1 px-2 flex flex-row">
-                            <FileHeader
-                                fileName={path}
-                                repo={{
-                                    name: repoInfo.name,
-                                    displayName: repoInfo.displayName,
-                                    webUrl: repoInfo.webUrl,
-                                    codeHostType: repoInfo.codeHostType,
-                                }}
-                                branchDisplayName={revisionName}
-                            />
-                        </div>
-                        <Separator />
-                    </>
-                )}
-            </div>
-            {isServiceError(repoInfo) ? (
-                <div className="flex h-full">
-                    <div className="m-auto flex flex-col items-center gap-2">
-                        <LuBookX className="h-12 w-12 text-secondary-foreground" />
-                        <span className="font-medium text-secondary-foreground">Repository not found</span>
-                    </div>
-                </div>
-            ) : (
-                <CodePreviewWrapper
-                    path={path}
-                    repoName={repoInfo.name}
-                    revisionName={revisionName ?? 'HEAD'}
-                    domain={params.domain}
-                />
-            )}
-        </div>
-    )
-}
-
-interface CodePreviewWrapper {
-    path: string,
-    repoName: string,
-    revisionName: string,
-    domain: string,
-}
-
-const CodePreviewWrapper = async ({
-    path,
-    repoName,
-    revisionName,
-    domain,
-}: CodePreviewWrapper) => {
-    // @todo: this will depend on `pathType`.
     const fileSourceResponse = await getFileSource({
         fileName: path,
         repository: repoName,
-        branch: revisionName,
-    }, domain);
+        branch: revisionName ?? 'HEAD',
+    }, params.domain);
 
     if (isServiceError(fileSourceResponse)) {
         if (fileSourceResponse.errorCode === ErrorCode.FILE_NOT_FOUND) {
@@ -142,13 +95,57 @@ const CodePreviewWrapper = async ({
         throw new ServiceErrorException(fileSourceResponse);
     }
 
+    const codeHostInfo = getCodeHostInfoForRepo({
+        codeHostType: repoInfo.codeHostType,
+        name: repoInfo.name,
+        displayName: repoInfo.displayName,
+        webUrl: repoInfo.webUrl,
+    });
+
     return (
-        <CodePreview
-            source={base64Decode(fileSourceResponse.source)}
-            language={fileSourceResponse.language}
-            repoName={repoName}
-            path={path}
-            revisionName={revisionName}
-        />
+        <>
+            <div className='sticky top-0 left-0 right-0 z-10'>
+                <TopBar
+                    defaultSearchQuery={`repo:${repoName}${revisionName ? ` rev:${revisionName}` : ''} `}
+                    domain={params.domain}
+                />
+                <Separator />
+                <div className="bg-accent py-1 px-2 flex flex-row items-center">
+                    <FileHeader
+                        fileName={path}
+                        repo={{
+                            name: repoInfo.name,
+                            displayName: repoInfo.displayName,
+                            webUrl: repoInfo.webUrl,
+                            codeHostType: repoInfo.codeHostType,
+                        }}
+                        branchDisplayName={revisionName}
+                    />
+                    {(fileSourceResponse.webUrl && codeHostInfo) && (
+                        <a
+                            href={fileSourceResponse.webUrl}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="flex flex-row items-center gap-2 px-2 py-0.5 rounded-md flex-shrink-0"
+                        >
+                            <Image
+                                src={codeHostInfo.icon}
+                                alt={codeHostInfo.codeHostName}
+                                className={cn('w-4 h-4 flex-shrink-0', codeHostInfo.iconClassName)}
+                            />
+                            <span className="text-sm font-medium">Open in {codeHostInfo.codeHostName}</span>
+                        </a>
+                    )}
+                </div>
+                <Separator />
+            </div>
+            <CodePreviewPanel
+                source={base64Decode(fileSourceResponse.source)}
+                language={fileSourceResponse.language}
+                repoName={repoInfo.name}
+                path={path}
+                revisionName={revisionName ?? 'HEAD'}
+            />
+        </>
     )
 }