+++ packages/web/src/app/[domain]/browse/hooks/useBrowseState.ts
@@ -1,0 +1,12 @@
+'use client';
+
+import { useContext } from "react";
+import { BrowseStateContext } from "../browseStateProvider";
+
+export const useBrowseState = () => {
+    const context = useContext(BrowseStateContext);
+    if (!context) {
+        throw new Error('useBrowseState must be used within a BrowseStateProvider');
+    }
+    return context;
+}