+++ packages/web/src/features/entitlements/constants.ts
@@ -17,6 +17,7 @@ const entitlements = [
     "public-access",
     "multi-tenancy",
     "sso",
+    "code-nav"
 ] as const;
 export type Entitlement = (typeof entitlements)[number];
 
@@ -27,7 +28,7 @@ export const isValidEntitlement = (entitlement: string): entitlement is Entitlem
 export const entitlementsByPlan: Record<Plan, Entitlement[]> = {
     oss: [],
     "cloud:team": ["billing", "multi-tenancy", "sso"],
-    "self-hosted:enterprise": ["search-contexts", "sso"],
-    "self-hosted:enterprise-unlimited": ["search-contexts", "public-access", "sso"],
+    "self-hosted:enterprise": ["search-contexts", "sso", "code-nav"],
+    "self-hosted:enterprise-unlimited": ["search-contexts", "public-access", "sso", "code-nav"],
     "self-hosted:enterprise-custom": [],
 } as const;