+++ packages/web/src/features/search/searchApi.ts
@@ -6,7 +6,7 @@ import { prisma } from "@/prisma";
 import { ErrorCode } from "../../lib/errorCodes";
 import { StatusCodes } from "http-status-codes";
 import { zoektSearchResponseSchema } from "./zoektSchema";
-import { SearchRequest, SearchResponse, SearchResultRange } from "./types";
+import { SearchRequest, SearchResponse, SourceRange } from "./types";
 import { OrgRole, Repo } from "@sourcebot/db";
 import * as Sentry from "@sentry/nextjs";
 import { sew, withAuth, withOrgMembership } from "@/actions";
@@ -213,6 +213,92 @@ export const search = async ({ query, matches, contextLines, whole }: SearchRequ
                     }
                 })).forEach(repo => repos.set(repo.name, repo));
 
+                const files = Result.Files?.map((file) => {
+                    const fileNameChunks = file.ChunkMatches.filter((chunk) => chunk.FileName);
+
+                    const webUrl = (() => {
+                        const template: string | undefined = Result.RepoURLs[file.Repository];
+                        if (!template) {
+                            return undefined;
+                        }
+
+                        // If there are multiple branches pointing to the same revision of this file, it doesn't
+                        // matter which branch we use here, so use the first one.
+                        const branch = file.Branches && file.Branches.length > 0 ? file.Branches[0] : "HEAD";
+                        return getFileWebUrl(template, branch, file.FileName);
+                    })();
+
+                    const identifier = file.RepositoryID ?? file.Repository;
+                    const repo = repos.get(identifier);
+
+                    // This should never happen... but if it does, we skip the file.
+                    if (!repo) {
+                        Sentry.captureMessage(
+                            `Repository not found for identifier: ${identifier}; skipping file "${file.FileName}"`,
+                            'warning'
+                        );
+                        return undefined;
+                    }
+
+                    return {
+                        fileName: {
+                            text: file.FileName,
+                            matchRanges: fileNameChunks.length === 1 ? fileNameChunks[0].Ranges.map((range) => ({
+                                start: {
+                                    byteOffset: range.Start.ByteOffset,
+                                    column: range.Start.Column,
+                                    lineNumber: range.Start.LineNumber,
+                                },
+                                end: {
+                                    byteOffset: range.End.ByteOffset,
+                                    column: range.End.Column,
+                                    lineNumber: range.End.LineNumber,
+                                }
+                            })) : [],
+                        },
+                        repository: repo.name,
+                        repositoryId: repo.id,
+                        webUrl: webUrl,
+                        language: file.Language,
+                        chunks: file.ChunkMatches
+                            .filter((chunk) => !chunk.FileName) // Filter out filename chunks.
+                            .map((chunk) => {
+                                return {
+                                    content: chunk.Content,
+                                    matchRanges: chunk.Ranges.map((range) => ({
+                                        start: {
+                                            byteOffset: range.Start.ByteOffset,
+                                            column: range.Start.Column,
+                                            lineNumber: range.Start.LineNumber,
+                                        },
+                                        end: {
+                                            byteOffset: range.End.ByteOffset,
+                                            column: range.End.Column,
+                                            lineNumber: range.End.LineNumber,
+                                        }
+                                    }) satisfies SourceRange),
+                                    contentStart: {
+                                        byteOffset: chunk.ContentStart.ByteOffset,
+                                        column: chunk.ContentStart.Column,
+                                        lineNumber: chunk.ContentStart.LineNumber,
+                                    },
+                                    symbols: chunk.SymbolInfo?.map((symbol) => {
+                                        return {
+                                            symbol: symbol.Sym,
+                                            kind: symbol.Kind,
+                                            parent: symbol.Parent.length > 0 ? {
+                                                symbol: symbol.Parent,
+                                                kind: symbol.ParentKind,
+                                            } : undefined,
+                                        }
+                                    }) ?? undefined,
+                                }
+                            }),
+                        branches: file.Branches,
+                        content: file.Content,
+                    }
+                }).filter((file) => file !== undefined) ?? [];
+
                 return {
                     zoektStats: {
                         duration: Result.Duration,
@@ -236,91 +322,7 @@ export const search = async ({ query, matches, contextLines, whole }: SearchRequ
                         regexpsConsidered: Result.RegexpsConsidered,
                         flushReason: Result.FlushReason,
                     },
-                    files: Result.Files?.map((file) => {
-                        const fileNameChunks = file.ChunkMatches.filter((chunk) => chunk.FileName);
-
-                        const webUrl = (() => {
-                            const template: string | undefined = Result.RepoURLs[file.Repository];
-                            if (!template) {
-                                return undefined;
-                            }
-
-                            // If there are multiple branches pointing to the same revision of this file, it doesn't
-                            // matter which branch we use here, so use the first one.
-                            const branch = file.Branches && file.Branches.length > 0 ? file.Branches[0] : "HEAD";
-                            return getFileWebUrl(template, branch, file.FileName);
-                        })();
-
-                        const identifier = file.RepositoryID ?? file.Repository;
-                        const repo = repos.get(identifier);
-
-                        // This should never happen... but if it does, we skip the file.
-                        if (!repo) {
-                            Sentry.captureMessage(
-                                `Repository not found for identifier: ${identifier}; skipping file "${file.FileName}"`,
-                                'warning'
-                            );
-                            return undefined;
-                        }
-
-                        return {
-                            fileName: {
-                                text: file.FileName,
-                                matchRanges: fileNameChunks.length === 1 ? fileNameChunks[0].Ranges.map((range) => ({
-                                    start: {
-                                        byteOffset: range.Start.ByteOffset,
-                                        column: range.Start.Column,
-                                        lineNumber: range.Start.LineNumber,
-                                    },
-                                    end: {
-                                        byteOffset: range.End.ByteOffset,
-                                        column: range.End.Column,
-                                        lineNumber: range.End.LineNumber,
-                                    }
-                                })) : [],
-                            },
-                            repository: repo.name,
-                            repositoryId: repo.id,
-                            webUrl: webUrl,
-                            language: file.Language,
-                            chunks: file.ChunkMatches
-                                .filter((chunk) => !chunk.FileName) // Filter out filename chunks.
-                                .map((chunk) => {
-                                    return {
-                                        content: chunk.Content,
-                                        matchRanges: chunk.Ranges.map((range) => ({
-                                            start: {
-                                                byteOffset: range.Start.ByteOffset,
-                                                column: range.Start.Column,
-                                                lineNumber: range.Start.LineNumber,
-                                            },
-                                            end: {
-                                                byteOffset: range.End.ByteOffset,
-                                                column: range.End.Column,
-                                                lineNumber: range.End.LineNumber,
-                                            }
-                                        }) satisfies SearchResultRange),
-                                        contentStart: {
-                                            byteOffset: chunk.ContentStart.ByteOffset,
-                                            column: chunk.ContentStart.Column,
-                                            lineNumber: chunk.ContentStart.LineNumber,
-                                        },
-                                        symbols: chunk.SymbolInfo?.map((symbol) => {
-                                            return {
-                                                symbol: symbol.Sym,
-                                                kind: symbol.Kind,
-                                                parent: symbol.Parent.length > 0 ? {
-                                                    symbol: symbol.Parent,
-                                                    kind: symbol.ParentKind,
-                                                } : undefined,
-                                            }
-                                        }) ?? undefined,
-                                    }
-                                }),
-                            branches: file.Branches,
-                            content: file.Content,
-                        }
-                    }).filter((file) => file !== undefined) ?? [],
+                    files,
                     repositoryInfo: Array.from(repos.values()).map((repo) => ({
                         id: repo.id,
                         codeHostType: repo.external_codeHostType,
@@ -329,6 +331,16 @@ export const search = async ({ query, matches, contextLines, whole }: SearchRequ
                         webUrl: repo.webUrl ?? undefined,
                     })),
                     isBranchFilteringEnabled: isBranchFilteringEnabled,
+                    stats: {
+                        matchCount: files.reduce(
+                            (acc, file) =>
+                                acc + file.chunks.reduce(
+                                    (acc, chunk) => acc + chunk.matchRanges.length,
+                                    0,
+                                ),
+                            0,
+                        )
+                    }
                 } satisfies SearchResponse;
             });